cleaned up TDD Unit Tests
